+++ R/CubeRoot.R
@@ -1,4 +1,7 @@
-# Function that computes the (real) cube root of a number
+#'@title Compute the non negative cube root of a real number 
+#'
+#'@param x a non negative number
+#'
 CubeRoot <- function(x){
   
   sign(x)*abs(x)^(1/3)