+++ R/Q3.R
@@ -1,0 +1,7 @@
+#'@title Computes the third quartile
+#'
+#'@param x vector of real numbers
+#'
+Q3 <- function(x) {
+  return(quantile(x , probs = c(.75)))
+}