+++ R/quiet.R
@@ -1,5 +1,7 @@
-## Function to prevent another function from printing things as part of its execution
-
+#'@title Function to prevent another function from printing things as part of its execution
+#'
+#'@param x code 
+#'
 quiet <- function(x) { 
   sink(tempfile()) 
   on.exit(sink())