+++ NEWS.md
@@ -1,8 +1,10 @@
 # traj 3.1.1
 
 - Added the extractor functions `trajPartition()`, `trajFuzzyPartition()`, `trajClusterSummary()` and `trajAnalaysis()` acting on a `trajClusters` object to extract respectively the partition, the fuzzy partition, the cluster-wise summaries of the measures and an order analysis on the cluster-wise medians.
+- Added the extractor functions `trajReducedPartition()`, `trajReducedFuzzyPartition()` acting on a `trajReduce` object to extract respectively the reduced partition  and the reduced fuzzy partition.
 - Simplified the structure and output of `summary()` acting on a `trajClusters` object.
-- Renamed the `scatterplots()` function to `trajScatter()` for uniformity
+- Renamed the `scatterplots()` function to `trajScatter()` for uniformity.
+- Added documentation for all the internal functions as well as for `CVIplot()` and `trajScatter()`.
 
 # traj 3.1.0